Implement fixes for all security audit findings
Remediates every Critical/High/Medium/Low finding from SECURITY_AUDIT.md. Full
test suite green: 374 tests across all modules.

Consensus (C1, H2, L15) — no reward/work minted without matching work:
- checkLeadingZeroBits rejects a non-positive challenge (difficulty 0 no longer
  passes PoW with zero work); registerOrphan refuses sub-minDifficulty blocks.
- uncleEligible/validateUncles/selectUncles and HeaderChain.uncleWork bound each
  uncle's difficulty to [minDifficulty, nephew difficulty] via one shared rule, so
  a flat uncle reward can't be minted for near-zero work and the headers-first
  work gate can't be inflated toward 2^255. ChainSynchronizer folds in the same
  bounded uncle work.

Contract VM (C2, H4, M3, L12) — deterministic, no JVM-stack-dependent fork:
- DepthLimitedInterpreterMachine caps WASM call depth at a fixed 1024 via Chicory's
  own callStack; execution runs on a fixed 64 MiB stack thread, so the trip point
  is a network constant and depth-exceeded is a deterministic revert.
- Modules are parsed/validated once and cached by code hash; deploy validates and
  size-caps code (256 KiB). Float rejection now covers SIMD/vector-float opcodes.
  box_read charges for the full serialized box size.

Dependency (H1): BouncyCastle 1.76 -> 1.78.1 (CVE-2024-30172 Ed25519 DoS).

P2P (H3, M1, L1, L2): /add_peer offloads DNS off the event-loop; PeerBroadcaster
pins each send (anti-rebinding SSRF); bans mirror onto a hostname key; the SSRF
filter defaults on for any non-loopback-advertised node.

Wallet (H5, L4): key file written owner-only (0600) atomically; PBKDF2 200k->600k.

Node API (M2, L3, L9): weighted rate limiting for deep scans / dry-runs and lower
SCAN_DEPTH_MAX; generic client error messages (detail logged server-side); bounded
ScanPredicate nesting.

Core robustness (M4, M5, L5, L6, L7, L8, L10, L13, L14): signature check fails
closed on an empty key; MerkleTree gains leaf/node domain separation, canonical
construction and a defined empty root; optional mempool min-fee; JSON tx amounts
read as long; SnapshotChunk bounds; RIPEMD160Hash.toBytes; bounded FloodDiscovery.

CI (I1): full-build+test+dependency-review workflow provided as
.github/ci-workflow.yml.example (the automation account lacks GitHub `workflows`
permission to install it under .github/workflows/ directly).

Tests updated for the new merkle scheme and made deterministic (no flaky
random-nonce PoW check); added WasmDepthLimitTest.

+++ lib-core/src/main/java/rhizome/core/transaction/Transaction.java
@@ -224,7 +224,7 @@ public sealed interface Transaction permits TransactionImpl {
         public Transaction fromJson(JSONObject json) {
             var builder = TransactionImpl.builder()
                 .timestamp(json.getLong(TIMESTAMP))
-                .fee(new TransactionAmount(json.getInt(FEE)))
+                .fee(new TransactionAmount(json.getLong(FEE)))
                 .chainId(json.optInt(CHAIN_ID, 0))
                 .nonce(json.optLong(NONCE, 0))
                 .to(PublicAddress.of(json.getString(TO)));
@@ -238,12 +238,12 @@ public sealed interface Transaction permits TransactionImpl {
 
         
             if (json.getString("from").isEmpty()) {
-                builder.amount(new TransactionAmount(json.getInt(AMOUNT)))
+                builder.amount(new TransactionAmount(json.getLong(AMOUNT)))
                     .isTransactionFee(true);
             } else {
                 builder.from(PublicAddress.of(json.getString(FROM)))
                     .signature(TransactionSignature.of(json.getString(SIGNATURE)))
-                    .amount(new TransactionAmount(json.getInt(AMOUNT)))
+                    .amount(new TransactionAmount(json.getLong(AMOUNT)))
                     .isTransactionFee(false)
                     .signingKey(PublicKey.of(json.getString(SIGNING_KEY)));
             }
